+++ Package.swift
@@ -7,7 +7,7 @@ let package = Package(
         .library(name: "Chroma", targets: ["Chroma"]),
     ],
     dependencies: [
-        .package(url: "https://github.com/SolarDesignSystem/Matrix-Swift.git", from: "1.0.5"),
+        .package(url: "https://github.com/SolarDesignSystem/Matrix-Swift.git", from: "1.0.6"),
         .package(url: "https://github.com/apple/swift-docc-plugin", from: "1.0.0"),
     ],
     targets: [