+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added 
+
+- Added support for the RGB color space and its various profiles.
+
 ## [1.0.1] - 2023-04-06
 
 ### Fixed